Add more natural camera controls
Now the camera moves alongside default x, y and z axes, considering only
rotations on the z axis, thus providing a more natural navigation
experience. Moving back and forth only moves it on a z-rotated x axis;
moving it right and left does so on a z-rotated y axis; moving it up and
down does so in the world's default "up" direction.